refactor add_row_area_of_effect
